refactor : BigDecimal로 변경 #24

+++ src/main/kotlin/com/get_offer/auction/controller/BidRequest.kt
@@ -1,5 +1,7 @@
 package com.get_offer.auction.controller
 
+import java.math.BigDecimal
+
 data class BidRequest(
-    val bidPrice: Int
+    val bidPrice: BigDecimal
 )
